+++ src/main/java/micdoodle8/mods/galacticraft/core/client/model/ModelParaChest.java
@@ -76,31 +76,31 @@ public class ModelParaChest extends ChestModel
             this.parachuteStrings[i].render(0.0625F);
         }
 
-        this.parachute[0].rotateAngleY = (float) (0 / Constants.RADIANS_TO_DEGREES);
-        this.parachute[2].rotateAngleY = (float) -(0 / Constants.RADIANS_TO_DEGREES);
-        this.parachuteStrings[0].rotateAngleY = (float) (0 / Constants.RADIANS_TO_DEGREES);
-        this.parachuteStrings[1].rotateAngleY = (float) (0 / Constants.RADIANS_TO_DEGREES);
-        this.parachuteStrings[2].rotateAngleY = (float) -(0 / Constants.RADIANS_TO_DEGREES);
-        this.parachuteStrings[3].rotateAngleY = (float) -(0 / Constants.RADIANS_TO_DEGREES);
+        this.parachute[0].rotateAngleY = 0 / Constants.RADIANS_TO_DEGREES;
+        this.parachute[2].rotateAngleY = -(0 / Constants.RADIANS_TO_DEGREES);
+        this.parachuteStrings[0].rotateAngleY = 0 / Constants.RADIANS_TO_DEGREES;
+        this.parachuteStrings[1].rotateAngleY = 0 / Constants.RADIANS_TO_DEGREES;
+        this.parachuteStrings[2].rotateAngleY = -(0 / Constants.RADIANS_TO_DEGREES);
+        this.parachuteStrings[3].rotateAngleY = -(0 / Constants.RADIANS_TO_DEGREES);
 
         this.parachute[0].setRotationPoint(-5.85F, -11.0F, 2.0F);
         this.parachute[1].setRotationPoint(9F, -7F, 2.0F);
         this.parachute[2].setRotationPoint(-2.15F, 4.0F, 2.0F);
-        this.parachute[0].rotateAngleZ = (float) (210F / Constants.RADIANS_TO_DEGREES);
-        this.parachute[1].rotateAngleZ = (float) (180F / Constants.RADIANS_TO_DEGREES);
-        this.parachute[2].rotateAngleZ = (float) -(210F / Constants.RADIANS_TO_DEGREES);
-        this.parachuteStrings[0].rotateAngleZ = (float) ((155F + 180F) / Constants.RADIANS_TO_DEGREES);
-        this.parachuteStrings[0].rotateAngleX = (float) (23F / Constants.RADIANS_TO_DEGREES);
+        this.parachute[0].rotateAngleZ = 210F / Constants.RADIANS_TO_DEGREES;
+        this.parachute[1].rotateAngleZ = 180F / Constants.RADIANS_TO_DEGREES;
+        this.parachute[2].rotateAngleZ = -(210F / Constants.RADIANS_TO_DEGREES);
+        this.parachuteStrings[0].rotateAngleZ = (155F + 180F) / Constants.RADIANS_TO_DEGREES;
+        this.parachuteStrings[0].rotateAngleX = 23F / Constants.RADIANS_TO_DEGREES;
         this.parachuteStrings[0].setRotationPoint(9.0F, 3.0F, 2.0F);
-        this.parachuteStrings[1].rotateAngleZ = (float) ((155F + 180F) / Constants.RADIANS_TO_DEGREES);
-        this.parachuteStrings[1].rotateAngleX = (float) -(23F / Constants.RADIANS_TO_DEGREES);
+        this.parachuteStrings[1].rotateAngleZ = (155F + 180F) / Constants.RADIANS_TO_DEGREES;
+        this.parachuteStrings[1].rotateAngleX = -(23F / Constants.RADIANS_TO_DEGREES);
         this.parachuteStrings[1].setRotationPoint(9.0F, 3.0F, 2.0F);
 
-        this.parachuteStrings[2].rotateAngleZ = (float) -((155F + 180F) / Constants.RADIANS_TO_DEGREES);
-        this.parachuteStrings[2].rotateAngleX = (float) (23F / Constants.RADIANS_TO_DEGREES);
+        this.parachuteStrings[2].rotateAngleZ = -((155F + 180F) / Constants.RADIANS_TO_DEGREES);
+        this.parachuteStrings[2].rotateAngleX = 23F / Constants.RADIANS_TO_DEGREES;
         this.parachuteStrings[2].setRotationPoint(9.0F, 3.0F, 2.0F);
-        this.parachuteStrings[3].rotateAngleZ = (float) -((155F + 180F) / Constants.RADIANS_TO_DEGREES);
-        this.parachuteStrings[3].rotateAngleX = (float) -(23F / Constants.RADIANS_TO_DEGREES);
+        this.parachuteStrings[3].rotateAngleZ = -((155F + 180F) / Constants.RADIANS_TO_DEGREES);
+        this.parachuteStrings[3].rotateAngleX = -(23F / Constants.RADIANS_TO_DEGREES);
         this.parachuteStrings[3].setRotationPoint(9.0F, 3.0F, 2.0F);
 
         Minecraft.getInstance().textureManager.bindTexture(RenderParaChest.PARACHEST_TEXTURE);
