+++ src/main/java/micdoodle8/mods/galacticraft/core/client/render/entities/RenderSchematic.java
@@ -46,11 +46,11 @@ public class RenderSchematic extends EntityRenderer<EntityHangingSchematic>
         GlStateManager.popMatrix();
         super.doRender(entity, x, y, z, entityYaw, partialTicks);
     }
-    
+
     private void renderPainting(EntityHangingSchematic painting, int width, int height)
     {
-        float f = (float)(-width) / 2.0F;
-        float f1 = (float)(-height) / 2.0F;
+        float f = (float) (-width) / 2.0F;
+        float f1 = (float) (-height) / 2.0F;
         float f2 = 0.5F;
         float f3 = 0.75F;
         float f4 = 0.8125F;
@@ -69,15 +69,15 @@ public class RenderSchematic extends EntityRenderer<EntityHangingSchematic>
         {
             for (int j = 0; j < height / 16; ++j)
             {
-                double a = f + (float)((i + 1) * 16);
-                double b = f + (float)(i * 16);
-                double c = f1 + (float)((j + 1) * 16);
-                double d = f1 + (float)(j * 16);
+                double a = f + (float) ((i + 1) * 16);
+                double b = f + (float) (i * 16);
+                double c = f1 + (float) ((j + 1) * 16);
+                double d = f1 + (float) (j * 16);
                 this.setLightmap(painting, (a + b) / 2.0F, (c + d) / 2.0F);
-                float f19 = (float)(width - i * 16) / 32.0F;
-                float f20 = (float)(width - (i + 1) * 16) / 32.0F;
-                float f21 = (float)(height - j * 16) / 32.0F;
-                float f22 = (float)(height - (j + 1) * 16) / 32.0F;
+                float f19 = (float) (width - i * 16) / 32.0F;
+                float f20 = (float) (width - (i + 1) * 16) / 32.0F;
+                float f21 = (float) (height - j * 16) / 32.0F;
+                float f22 = (float) (height - (j + 1) * 16) / 32.0F;
                 Tessellator tessellator = Tessellator.getInstance();
                 BufferBuilder worldrenderer = tessellator.getBuffer();
                 worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX_NORMAL);
@@ -113,7 +113,7 @@ public class RenderSchematic extends EntityRenderer<EntityHangingSchematic>
     private void setLightmap(HangingEntity painting, double p_77008_2_, double p_77008_3_)
     {
         int i = MathHelper.floor(painting.posX);
-        int j = MathHelper.floor(painting.posY + (double)(p_77008_3_ / 16.0F));
+        int j = MathHelper.floor(painting.posY + (p_77008_3_ / 16.0F));
         int k = MathHelper.floor(painting.posZ);
         Direction enumfacing = painting.getHorizontalFacing();
 
@@ -140,7 +140,7 @@ public class RenderSchematic extends EntityRenderer<EntityHangingSchematic>
         int l = this.renderManager.world.getCombinedLight(new BlockPos(i, j, k), 0);
         int i1 = l % 65536;
         int j1 = l / 65536;
-        GLX.glMultiTexCoord2f(GLX.GL_TEXTURE1, (float)j1, (float)j1);
+        GLX.glMultiTexCoord2f(GLX.GL_TEXTURE1, (float) j1, (float) j1);
         GlStateManager.color3f(1.0F, 1.0F, 1.0F);
     }
 }
