+++ src/main/java/micdoodle8/mods/galacticraft/planets/venus/client/WeatherRendererVenus.java
@@ -30,10 +30,10 @@ public class WeatherRendererVenus implements IRenderHandler
     {
         for (int i = 0; i < 32; ++i)
         {
-            float f1 = (float)(i - 16);
+            float f1 = (float) (i - 16);
             for (int j = 0; j < 32; ++j)
             {
-                float f = (float)(j - 16);
+                float f = (float) (j - 16);
                 float f2 = MathHelper.sqrt(f * f + f1 * f1);
                 this.rainXCoords[i << 5 | j] = -f1 / f2;
                 this.rainYCoords[i << 5 | j] = f / f2;
@@ -49,11 +49,15 @@ public class WeatherRendererVenus implements IRenderHandler
         if (strength > 0.0F)
         {
             int rendererUpdateCount = 0;
-            try {
+            try
+            {
                 Field fieldRUC = mc.gameRenderer.getClass().getDeclaredField(GCCoreUtil.isDeobfuscated() ? "rendererUpdateCount" : "field_78529_t");
                 fieldRUC.setAccessible(true);
                 rendererUpdateCount = fieldRUC.getInt(mc.gameRenderer);
-            } catch (Exception e) { }
+            }
+            catch (Exception e)
+            {
+            }
             mc.gameRenderer.enableLightmap();
             Entity entity = mc.getRenderViewEntity();
 
@@ -64,9 +68,9 @@ public class WeatherRendererVenus implements IRenderHandler
             GlStateManager.enableBlend();
             GlStateManager.blendFuncSeparate(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA, GlStateManager.SourceFactor.ONE, GlStateManager.DestFactor.ZERO);
             GlStateManager.alphaFunc(516, 0.1F);
-            double d0 = entity.lastTickPosX + (entity.posX - entity.lastTickPosX) * (double)partialTicks;
-            double d1 = entity.lastTickPosY + (entity.posY - entity.lastTickPosY) * (double)partialTicks;
-            double d2 = entity.lastTickPosZ + (entity.posZ - entity.lastTickPosZ) * (double)partialTicks;
+            double d0 = entity.lastTickPosX + (entity.posX - entity.lastTickPosX) * (double) partialTicks;
+            double d1 = entity.lastTickPosY + (entity.posY - entity.lastTickPosY) * (double) partialTicks;
+            double d2 = entity.lastTickPosZ + (entity.posZ - entity.lastTickPosZ) * (double) partialTicks;
             int l = MathHelper.floor(d1);
 
             int r = 4;
@@ -90,7 +94,7 @@ public class WeatherRendererVenus implements IRenderHandler
                 for (int x = px - r; x <= px + r; ++x)
                 {
                     mutablePos.setPos(x, 0, z);
-                    int yHeight = world.getHeight(Heightmap.Type.MOTION_BLOCKING, mutablePos).getY() + 4 - (int)(4.8F * strength);
+                    int yHeight = world.getHeight(Heightmap.Type.MOTION_BLOCKING, mutablePos).getY() + 4 - (int) (4.8F * strength);
                     int y = py - r;
                     int ymax = py + r;
 
@@ -113,7 +117,7 @@ public class WeatherRendererVenus implements IRenderHandler
 
                     if (y != ymax)
                     {
-                        this.random.setSeed((long)(x * x * 3121 + x * 45238971 ^ z * z * 418711 + z * 13761));
+                        this.random.setSeed(x * x * 3121 + x * 45238971 ^ z * z * 418711 + z * 13761);
 
                         if (drawFlag != 0)
                         {
@@ -128,9 +132,9 @@ public class WeatherRendererVenus implements IRenderHandler
                         }
 
                         int index = indexZ + x - px + 16;
-                        double dx = (double)this.rainXCoords[index] * 0.5D;
-                        double dz = (double)this.rainYCoords[index] * 0.5D;
-                        double dy = -((double)(rendererUpdateCount + x * x * 3121 + x * 45238971 + z * z * 418711 + z * 13761 & 31) + (double)partialTicks) / 80.0D * (3.0D + this.random.nextDouble());
+                        double dx = (double) this.rainXCoords[index] * 0.5D;
+                        double dz = (double) this.rainYCoords[index] * 0.5D;
+                        double dy = -((double) (rendererUpdateCount + x * x * 3121 + x * 45238971 + z * z * 418711 + z * 13761 & 31) + (double) partialTicks) / 80.0D * (3.0D + this.random.nextDouble());
                         double yo = this.random.nextDouble() / 1.8D;
                         double xx = x + 0.5D - entity.posX;
                         double zz = z + 0.5D - entity.posZ;
@@ -142,10 +146,10 @@ public class WeatherRendererVenus implements IRenderHandler
                         int ly = light & 65535;
                         double xc = x + 0.5D;
                         double zc = z + 0.5D;
-                        worldrenderer.pos(xc - dx, (double)ymax - yo, zc - dz).tex(0.0D, (double)y * 0.25D + dy).color(1.0F, 1.0F, 1.0F, alpha).lightmap(lx, ly).endVertex();
-                        worldrenderer.pos(xc + dx, (double)ymax - yo, zc + dz).tex(1.0D, (double)y * 0.25D + dy).color(1.0F, 1.0F, 1.0F, alpha).lightmap(lx, ly).endVertex();
-                        worldrenderer.pos(xc + dx, (double)y - yo, zc + dz).tex(1.0D, (double)ymax * 0.25D + dy).color(1.0F, 1.0F, 1.0F, alpha).lightmap(lx, ly).endVertex();
-                        worldrenderer.pos(xc - dx, (double)y - yo, zc - dz).tex(0.0D, (double)ymax * 0.25D + dy).color(1.0F, 1.0F, 1.0F, alpha).lightmap(lx, ly).endVertex();
+                        worldrenderer.pos(xc - dx, (double) ymax - yo, zc - dz).tex(0.0D, (double) y * 0.25D + dy).color(1.0F, 1.0F, 1.0F, alpha).lightmap(lx, ly).endVertex();
+                        worldrenderer.pos(xc + dx, (double) ymax - yo, zc + dz).tex(1.0D, (double) y * 0.25D + dy).color(1.0F, 1.0F, 1.0F, alpha).lightmap(lx, ly).endVertex();
+                        worldrenderer.pos(xc + dx, (double) y - yo, zc + dz).tex(1.0D, (double) ymax * 0.25D + dy).color(1.0F, 1.0F, 1.0F, alpha).lightmap(lx, ly).endVertex();
+                        worldrenderer.pos(xc - dx, (double) y - yo, zc - dz).tex(0.0D, (double) ymax * 0.25D + dy).color(1.0F, 1.0F, 1.0F, alpha).lightmap(lx, ly).endVertex();
                     }
                 }
             }
